+++ scripts/build-test-app.sh
@@ -1,10 +1,38 @@
 #!/usr/bin/env bash
 set -e
 
-PLATFORM=$([[ "${@#--android}" = "$@" ]] && echo "ios" || echo "android")
-TARGET=$([[ "${@#--device}" = "$@" ]] && echo "emulator" || echo "device")
 ROOT="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"/..
 CDV=$ROOT/node_modules/.bin/cordova
+
+PLATFORM=ios
+TARGET=emulator
+
+while :; do
+  case $1 in
+    --android)
+      PLATFORM=android
+      ;;
+    --browser)
+      PLATFORM=browser
+      ;;
+    --ios)
+      PLATFORM=ios
+      ;;
+    --device)
+      TARGET=device
+      ;;
+    --emulator)
+      TARGET=emulator
+      ;;
+    -?*)
+      printf 'WARN: Unknown option (ignored): %s\n' "$1" >&2
+      ;;
+    *)
+      break
+   esac
+
+   shift
+done
 
 rm -rf $ROOT/temp
 mkdir $ROOT/temp